TASK2_Unemployement_Analysis
UNEMPLOYMENT ANALYSIS WITH PYTHON
Problem Statement :
Unemployment is measured by the unemployment rate which is the number of people who are unemployed as a percentage of the total labour force. We have seen a sharp increase in the unemployment rate during Covid-19, so analyzing the unemployment rate can be a good data science project.